Q-n-A-chatbot-with-langchain/requirement.txt
langchain
langchain_community
streamlit
python-dotenv
langchain_ollama
langchain_groq
langchain-text-splitters
chromadb